rename sendgrid to django_sendgrid
sendgrid’s official library is names “sendgrid”
