+++ django_sendgrid/constants.py
