+++ django_sendgrid/models.py
@@ -12,7 +12,7 @@ from six import string_types
 
 from .signals import sendgrid_email_sent
 from .signals import sendgrid_event_recieved
-from sendgrid.constants import (
+from django_sendgrid.constants import (
 	ARGUMENT_DATA_TYPE_UNKNOWN,
 	ARGUMENT_DATA_TYPE_BOOLEAN,
 	ARGUMENT_DATA_TYPE_INTEGER,
@@ -21,7 +21,7 @@ from sendgrid.constants import (
 	ARGUMENT_DATA_TYPE_STRING,
 	UNIQUE_ARGS_STORED_FOR_EVENTS_WITHOUT_MESSAGE_ID,
 )
-from sendgrid.signals import sendgrid_email_sent
+from django_sendgrid.signals import sendgrid_email_sent
 
 MAX_CATEGORIES_PER_EMAIL_MESSAGE = 10
