+++ django_sendgrid/urls.py
@@ -9,9 +9,9 @@ from .views import listener
 
 
 urlpatterns = patterns('',
-	url(r"^events/$", "sendgrid.views.listener", name="sendgrid_post_event"),
+	url(r"^events/$", "django_sendgrid.views.listener", name="sendgrid_post_event"),
 	url(r"^messages/(?P<message_id>[-\w]+)/attachments/$",
-		"sendgrid.views.download_attachments",
+		"django_sendgrid.views.download_attachments",
 		name="sendgrid_download_attachments"
 	),
 )
