+++ example_project/django_sendgrid
@@ -1,0 +1,1 @@
+../django_sendgrid/
