+++ example_project/main/urls.py
@@ -1,7 +1,7 @@
-from django.conf.urls.defaults import patterns, include, url
+from django.conf.urls import patterns, include, url
 
 from main.views import send_simple_email
 
-urlpatterns = patterns('',
+urlpatterns = [
 	url(r'^$', send_simple_email),
-)
+]
